% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_tissuesubtype_sex.R
\name{plot_tissuesubtype_sex}
\alias{plot_tissuesubtype_sex}
\title{Plot Gtex data}
\usage{
plot_tissuesubtype_sex(se, gene, title)
}
\arguments{
\item{se}{SummarizedExperiment}

\item{gene}{An ensembl id of gene}

\item{title}{title of plot, by default it is the gene ENSEMBL}
}
\value{
No return. Plots are shown on screen immediately.
}
\description{
Given a SummarizedExperiment object and a gene ENSEMBL, this function plots the mean expression level of this gene across different tissue subtype and different sex
}
