% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/se_escapee.R
\docType{data}
\name{se_escapee}
\alias{se_escapee}
\title{Summarized experiment}
\format{
a SumaarizedExperiment
}
\source{
Gtex
}
\usage{
se_escapee
}
\description{
A summarized experiment containing all the meta data and expression data of escapees.
}
\keyword{datasets}
